requirements.txt
torch>=2.0.0
tensorflow-federated>=0.20.0
pysyft>=0.6.0
numpy>=1.21.0
scipy>=1.7.0
tqdm>=4.65.0
opacus>=1.0.0